Fix errors in sps30

+++ components/sps30/include/util.h
@@ -4,6 +4,7 @@
 extern "C" {
 #endif
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define CRC8_POLYNOMIAL             (0x31)
@@ -13,7 +14,7 @@ uint8_t sps30_calc_cksum(const uint8_t* data, uint16_t count);
 
 uint8_t sps30_is_data_valid(const uint8_t *data, uint8_t size);
 
-uint8_t sps30_bytes_to_data(const uint8_t *bytes, uint8_t size, uint8_t *data);
+size_t sps30_bytes_to_data(const uint8_t *bytes, uint8_t size, uint8_t *data);
 
 uint8_t *sps30_uint32_to_bytes(const uint32_t value);
